fix(poller): correctly allow cancellation of timeouts after the first one
The `stop` method cancels the timeout stored in `nextTimer` if it's set but we were only initializing it the first time we set the timeout, not for any subsequent ones. The `stopped` flag means that `callback` was not called, but this can still cause a memory leak and delay shutdown of the Node process.

Assigning each timeout to the nextTimer variable ensures that the clearTimeout call works correctly to invalidate the timeout that has been set.